pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegramautomation"
version = "0.1.0"
description = "Telegram dispatch agent with Google Sheets integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot==21.6",
    "google-api-python-client==2.149.0",
    "google-auth==2.35.0",
    "APScheduler==3.10.4",
    "python-dotenv==1.0.1",
    "pydantic==2.9.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
